requirements.txt
asgiref==3.6.0
attrs==22.2.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudinary==1.32.0
colorama==0.4.6
coreapi==2.3.1
coverage==7.2.3
cryptography==40.0.1
defusedxml==0.7.1
dj-database-url==1.3.0
dj-rest-auth==3.0.0
django-admin-interface==0.25.0
django-allauth==0.52.0
django-cloudinary-storage==0.3.0
drf-extensions==0.7.1
django-celery-beat==2.6.0
django-cacheops==7.0.2
django-ckeditor-5==0.2.13
django-configurations==2.4.1
django-cors-headers==3.14.0
django-dbbackup==4.1.0
django-request-logging==0.7.5
django-shortuuidfield==0.1.3
django-filter
django-seed==0.3.1
loguru==0.7.2
fcm-django==1.0.13
django-nose==1.4.7
django_redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
drf-spectacular-sidecar==2023.4.1
django-storages[boto3]==1.14.3
django-debug-toolbar==4.3.0
factory-boy==3.2.1
Faker==18.4.0
flower==2.0.1
gunicorn==20.1.0
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
jsonschema==4.17.3
micropython-fcntl==0.0.3
micropython-ffilib==0.1.2
mypy-extensions==1.0.0
nose==1.3.7
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.1
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
platformdirs==3.2.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
readtime==3.0.0
redis==3.5.3
requests==2.28.2
requests-oauthlib==1.3.1
sentry-sdk==2.3.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
watchdog==4.0.0
whitenoise[brotli]==6.4.0
zipp==3.15.0
django-phonenumber-field==7.3.0
phonenumberslite==8.13.38
